Cleaned up the manifests and added basic config files
for logstash and filebeat this has been subject to testing and playing.
They are still work in progress, but could be used as a basis for dev

The basic settings came from the projects augmented with documentation.
the modules have been copied from the filebeat `modules.d` dir and been
augmented while testing with a edge router altho been retried for now.